// web/src/app/core/api/generated/models/paged-response-of-submission-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { SubmissionResponse } from '../models/submission-response';
export interface PagedResponseOfSubmissionResponse {
  items?: Array<SubmissionResponse>;
  pageNumber?: (number | string);
  pageSize?: (number | string);
  totalCount?: (number | string);
  totalPages?: (number | string);
}